api/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
scrapy==2.6
azure-storage-blob==2.1.0
pyOpenSSL==22.0.0 
crochet